Refactor types to allow Array to specify `ParseableType`
